pyproject.toml
[tool.poetry]
name = "skypro_homework_27"
version = "0.1.0"
description = ""
authors = ["TatianaBelova333 <101121235+TatianaBelova333@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
